feat: Add ai/python federated content and data
