+++ apps/docs/.gitignore
@@ -47,6 +47,7 @@ public/docs/
 /content/guides/deployment/terraform.mdx
 /content/guides/deployment/terraform/tutorial.mdx
 /content/guides/deployment/ci/
+/content/guides/ai/python/
 
 # Downloaded TypeDoc dumps under spec/reference/<lib>/<ver>/. Regenerated by
 # `cd apps/docs/spec && make download.tsdoc.v2`. Hand-authored files in the
